Create P3693.cpp
I've just passed P3693 for the first time. This problem is a famous for simulation. I tried it in May 2021 but failed.
Well, actually I debuged with the data of test points that I didn't pass. It would be suffering to debug without knowing where the bug is, especially when the code was too long and hard to read.
qwq.